requirements.txt
kagglehub
numpy
pillow
tqdm
matplotlib
boto3
ipykernel
sagemaker>=2,<3

